refactor(compartment-mapper)!: Rearrange entry point modules
