+++ packages/compartment-mapper/archive.js
@@ -1,0 +1,1 @@
+export { makeArchive, writeArchive } from './src/archive.js';
