+++ packages/compartment-mapper/import-archive.js
@@ -1,0 +1,5 @@
+export {
+  parseArchive,
+  loadArchive,
+  importArchive,
+} from './src/import-archive.js';
