+++ packages/compartment-mapper/import.js
@@ -1,0 +1,1 @@
+export { loadLocation, importLocation } from './src/import.js';
