+++ packages/compartment-mapper/index.js
@@ -1,0 +1,11 @@
+// @ts-check
+export { loadLocation, importLocation } from './src/import.js';
+export { makeArchive, writeArchive } from './src/archive.js';
+export {
+  parseArchive,
+  loadArchive,
+  importArchive,
+} from './src/import-archive.js';
+export { search } from './src/search.js';
+export { compartmentMapForNodeModules } from './src/node-modules.js';
+export { makeBundle, writeBundle } from './src/bundle.js';
