--- packages/compartment-mapper/src/main.js
@@ -1,7 +1,0 @@
-// @ts-check
-export { loadLocation, importLocation } from './import.js';
-export { makeArchive, writeArchive } from './archive.js';
-export { parseArchive, loadArchive, importArchive } from './import-archive.js';
-export { search } from './search.js';
-export { compartmentMapForNodeModules } from './node-modules.js';
-export { makeBundle, writeBundle } from './bundle.js';
